requirements.txt
flask==2.3.3
gunicorn==21.2.0
langchain==0.2.0
langchain-core==0.3.0
langchain-community==0.2.0
langchain-huggingface==0.1.0
langchain-pinecone==0.0.9
langchain-groq==0.3.8
sentence-transformers==2.2.2
pypdf==3.11.0
pinecone-client==2.2.2
python-dotenv==1.0.0
torch==2.2.0
accelerate==1.23.0
requests==2.31.0